Add the Tokenizer trait to abstract over tokenizers
Also provide two implementations:

1. FiniteStateTokenizer: use a transducer without any pre- or
   postprocessing.
2. AlpinoTokenizer: use a transducer, applying the pre- and
   postprocessing needed for the Alpino tokenizer.
