+++ alpino-tokenizer/src/lib.rs
@@ -13,16 +13,22 @@
 //! use std::fs::File;
 //! use std::io::BufReader;
 //!
-//! use alpino_tokenizer::Tokenizer;
+//! use alpino_tokenizer::{AlpinoTokenizer, Tokenizer};
 //!
 //! let read = BufReader::new(File::open("testdata/toy.proto").unwrap());
-//! let tokenizer = Tokenizer::from_buf_read(read).unwrap();
+//! let tokenizer = AlpinoTokenizer::from_buf_read(read).unwrap();
 //!
 //! assert_eq!(
 //!   tokenizer.tokenize("Groningen is een Hanzestad. Groningen heeft veel bezienswaardigheden.").unwrap(),
 //!   vec![vec!["Groningen", "is", "een", "Hanzestad", "."],
 //!        vec!["Groningen", "heeft", "veel", "bezienswaardigheden", "."]]);
 //! ```
+
+mod alpino;
+pub use alpino::AlpinoTokenizer;
+
+mod fst;
+pub use fst::FiniteStateTokenizer;
 
 mod proto;
 
@@ -31,6 +37,6 @@ mod preproc;
 mod postproc;
 
 mod tokenizer;
-pub use tokenizer::Tokenizer;
+pub use tokenizer::{Tokenizer, TokenizerError};
 
 mod util;
